sync-github-releases: split tests to co-locate with their modules
index.spec.ts tested nothing in index.ts — it covered the other modules. Split it
into a spec beside each module (vike's own convention), so each file tests one
module: utils/changelog.spec.ts, utils/git.spec.ts, release-plan.spec.ts.
github.ts keeps no spec — getRepository is a trivial split and the rest is network
I/O not worth mocking here.
